requirements.txt
SpeechRecognition==3.14.3
pyttsx3==2.90
requests==2.32.3
python-dotenv==1.1.0
PyAudio==0.2.14 